+++ CHANGELOG.md
@@ -30,7 +30,8 @@
 
 # 2.1.0
 - 'EmTime' added logLevel in constructor
+- 'EmString new methods: startsWith, endsWith, indexOf, substring, getToken, isToken
 - 'EmStore' added logLevel in constructor
-- 'EmStore' added new methods: 'getStringLength', 'hasValue', 'hasBytes' and 'hasString'
+- 'EmString' added new methods: 'getStringLength', 'hasValue', 'hasBytes' and 'hasString'
 - Added new 'EmTag' & 'EmTags' classes to have simple synchronization among same values coming from different sources
 - Added new 'EmAutoPtr' class as a lightweight alternative to std::auto_ptr for environments where the standard library is not available or desired